Updated benchmarks and results
