+++ README.md
@@ -159,9 +159,15 @@ with open('/tmp/100.dat', 'w') as f:
 
 ## Benchmarks
 
-The benchmark does 1000 get requests against a local nginx server with
-a concurrency of 10. See *benchmarks* folder.
+The benchmark does 10000 get requests against a local nginx server in the default 
+configuration with a concurrency of 10. See `benchmarks` folder. The requests per
+second for a couple of popular clients is given in the table below.
 
-- httplib2 with geventhttpclient monkey patch (*benchmarks/httplib2_patched.py*): **~2500 req/s**
-- geventhttpclient.HTTPClient (*benchmarks/httpclient.py*): **~4000 req/s**
+| HTTP Client       | RPS    |
+|-------------------|--------|
+| GeventHTTPClient  | 7110.3 |
+| Urllib3           | 2276.1 |
+| Requests          | 1020.5 |
+| Httpx             | 737.7  |
 
+*Linux(x86_64), Python 3.11.6 @ Intel i7-7560U*
