+++ benchmarks/urllib3_bench.py
@@ -7,26 +7,27 @@ gevent.monkey.patch_all()
 
 import urllib3
 
-if __name__ == "__main__":
-    N = 1000
-    C = 10
+N = 1000
+C = 10
 
-    url = "http://127.0.0.1/index.html"
+url = "http://127.0.0.1/"
 
-    def run(client):
-        response = client.request("GET", url)
-        assert response.status == 200
 
-    client = urllib3.PoolManager()
-    group = gevent.pool.Pool(size=C)
+def run(client):
+    response = client.request("GET", url)
+    assert response.status == 200
 
-    for i in range(5):
-        now = time.time()
-        for _ in range(N):
-            group.spawn(run, client)
-        group.join()
 
-        delta = time.time() - now
-        req_per_sec = N / delta
+client = urllib3.PoolManager()
+group = gevent.pool.Pool(size=C)
 
-        print(f"request count:{N}, concurrenry:{C}, {req_per_sec} req/s")
+for i in range(5):
+    now = time.time()
+    for _ in range(N):
+        group.spawn(run, client)
+    group.join()
+
+    delta = time.time() - now
+    req_per_sec = N / delta
+
+    print(f"request count:{N}, concurrenry:{C}, {req_per_sec} req/s")
